backend/requirements.txt
alabaster==0.7.12
setuptools==68.0.0
amqp==5.0.6
aniso8601==9.0.1
argh==0.26.2
astroid==3.1.0
attrs==20.3.0
Babel==2.9.0
billiard==3.6.4.0
blueprint==3.4.2
celery==5.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.5
docutils
et-xmlfile==1.1.0
expiringdict==1.2.1
flake8==3.8.4
flake8-polyfill==1.0.2
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
html2text==2020.1.16
humanize==3.9.0
idna==2.10
imagesize==1.2.0
iniconfig==2.0.0
isort==5.6.4
itsdangerous==1.1.0
jenkinsapi==0.3.11
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.2.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.3.0
numpy==1.21.2
openpyxl==3.1.2
packaging==20.4
pandas==1.4.3
pathtools==0.1.2
pdfkit==0.6.1
pep257==0.7.0
pep8-naming==0.10.0
pkcs7==0.1.2
pluggy==1.4.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
py==1.11.0
py4j==0.10.9
pycodestyle==2.6.0
pycryptodome==3.9.9
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyspark==3.1.2; sys_platform != "Linux"
pytest==7.1.2
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
scipy==1.7.3
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stripe==5.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.26.2
uwsgitop==0.11
vine==5.0.0
watchdog==0.10.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
nameparser==1.1.3
python-jenkins==1.8.2
squareup==35.0.0.20240222
pyap==0.3.1
pycurl==7.43.0.6; sys_platform != "Linux"
googlemaps==4.10.0
boto3==1.16.21
botocore==1.19.21
s3transfer==0.3.3
jira==3.8.0
sendgrid==6.11.0
beautifulsoup4==4.11.1
